Fix warnings.

+++ tbrowser/HeaderBar.cpp
@@ -40,7 +40,7 @@ void HeaderBar::Redraw() const
     wprintw( m_win, " :: " );
     wattron( m_win, COLOR_PAIR(11) );
 
-    wprintw( m_win, "%.*s", m_archiveLen, m_archive );
+    wprintw( m_win, "%.*s", (int)m_archiveLen, m_archive );
 
     if( m_desc )
     {
@@ -50,7 +50,7 @@ void HeaderBar::Redraw() const
 
         int w = getmaxx( m_win ) - 23 - m_archiveLen;
         auto end = utfendcrlf( m_desc, w );
-        wprintw( m_win, "%.*s", end - m_desc, m_desc );
+        wprintw( m_win, "%.*s", int( end - m_desc ), m_desc );
     }
 
     if( m_galaxy )
